+++ test/cases/require-in-all-pug-references/webpack.config.js
@@ -1,0 +1,45 @@
+const path = require('path');
+const PugPlugin = require('../../pug-plugin');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '/',
+  },
+
+  entry: {
+    index: 'src/index.pug',
+  },
+
+  plugins: [new PugPlugin()],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: 'pug-loader',
+        options: {
+          method: 'render',
+          embedFilters: {
+            code: true,
+          },
+        },
+      },
+
+      {
+        test: /\.(css)$/,
+        use: ['css-loader'],
+      },
+
+      {
+        test: /\.(png|jpg|jpeg)/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};